pyproject.toml
[tool.poetry]
name = "statisticsapi"
version = "0.1.0"
description = ""
authors = ["whatislove118 <oleg.pochikalin@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
drf-spectacular = "^0.22.1"
django-filter = "^21.1"


[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"
autoflake = "^1.4"

[tool.pytest.ini_options]
addopts = "-v -s"
testpaths = [
    "tests",
]
pythonpath = [
    "."
]
python_files = "tests.py test_*.py *_tests.py tests/test_*.py"

[tool.black]
exclude = '''
/(
    .git
  | .mypy_cache
  | .tox
  | venv
  | .venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
# make it compatible with black
profile = "black" 

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
